+++ states/splash/LanguageSelectionScreen/assets/images/Flags/Spec/FlagSwedenSpec.c
@@ -20,8 +20,8 @@
 //												DECLARATIONS
 //---------------------------------------------------------------------------------------------------------
 
-extern BYTE FlagSwedenTiles[];
-extern BYTE FlagSwedenMap[];
+extern uint32 FlagSwedenTiles[];
+extern uint16 FlagSwedenMap[];
 
 
 //---------------------------------------------------------------------------------------------------------
