+++ states/splash/LanguageSelectionScreen/assets/images/Flags/Spec/FlagUnitedKingdomSpec.c
@@ -20,8 +20,8 @@
 //												DECLARATIONS
 //---------------------------------------------------------------------------------------------------------
 
-extern BYTE FlagUnitedKingdomTiles[];
-extern BYTE FlagUnitedKingdomMap[];
+extern uint32 FlagUnitedKingdomTiles[];
+extern uint16 FlagUnitedKingdomMap[];
 
 
 //---------------------------------------------------------------------------------------------------------
